Refactor global epsilon handling in DynamicalSystem and SoftRobot classes
- Moved global epsilon initialization and tangent epsilon property from DynamicalSystem to SoftRobot.
- Updated SoftRobot constructor to handle optional epsilon value for numerical computations.
- Enhanced documentation for global epsilon and tangent epsilon properties.
